Go back to custom array op compilers
